fix(chat): move runtime selector to separate field
