+++ apps/chat/src/components/Common/ApplicationWizard/CodeAppView/CodeAppView.tsx
@@ -50,6 +50,7 @@ import { CustomLogoSelect } from '@/src/components/Settings/CustomLogoSelect';
 
 import { ViewProps } from '../view-props';
 import { CodeEditor } from './CodeEditor';
+import { RuntimeVersionSelector } from './RuntimeVersionSelector';
 
 const features = [
   {
@@ -79,6 +80,7 @@ const MappingsForm = withLabel(
   DynamicFormFields<FormData, 'endpoints' | 'env'>,
 );
 const ComboBoxField = withErrorMessage(withLabel(MultipleComboBox));
+const RuntimeSelector = withController(withLabel(RuntimeVersionSelector));
 
 export const CodeAppView: React.FC<ViewProps> = ({
   onClose,
@@ -283,6 +285,12 @@ export const CodeAppView: React.FC<ViewProps> = ({
             selectedRuntime={watch('runtime')}
           />
         )}
+
+        <RuntimeSelector
+          control={control}
+          name="runtime"
+          label={t('Runtime version')}
+        />
 
         <MappingsForm
           label={t('Endpoints')}
